# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BattleReadyShelf API
  version: 0.1.0
  description: BattleReadyShelf API
servers:
  - url: /
paths:
...
  /api/v1/users/me:
    get:
      tags: [ "Users" ]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The current authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/v1/wargear-definitions:
    get:
      tags: [ "PersonalModelDefinitions" ]
      summary: List the wargear the signed-in user can attach to a personal model definition
      description: The shared catalogue plus any wargear this user created for their own definitions.
      operationId: getAvailableWargearDefinitions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The wargear available to this user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WargearDefinition'
  /api/v1/my/model-definitions:
    get:
      tags: [ "PersonalModelDefinitions" ]
      summary: List the signed-in user's own model definitions
      description: >-
        Both kinds of personal definition: customisations of a shared definition (which carry a
        baseModelDefinitionId) and ones the user wrote from scratch (which do not).
      operationId: getMyModelDefinitions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: This user's model definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDefinition'
    post:
      tags: [ "PersonalModelDefinitions" ]
      summary: Create a model definition of the user's own
      operationId: createMyModelDefinition
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertModelDefinitionDraftRequest'
      responses:
        '201':
          description: The created definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
  /api/v1/my/model-definitions/shared:
    get:
      tags: [ "PersonalModelDefinitions" ]
      summary: List the shared catalogue without the signed-in user's customisations
      description: >-
        GET /api/v1/model-definitions returns the shared catalogue and the user's own definitions
        together, so a customisation appears alongside the definition it was forked from. This
        endpoint returns the shared definitions on their own: to offer them for customisation, and
        to diff a personal copy against the definition it came from.
      operationId: getSharedModelDefinitions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The shared model definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDefinition'
  /api/v1/my/model-definitions/{modelDefinitionId}:
    put:
      tags: [ "PersonalModelDefinitions" ]
      summary: Replace the full contents of one of the user's own model definitions
      operationId: updateMyModelDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: modelDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertModelDefinitionDraftRequest'
      responses:
        '200':
          description: The updated definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
        '404':
          description: No such definition belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags: [ "PersonalModelDefinitions" ]
      summary: Delete one of the user's own model definitions, reverting to the shared one if it was a customisation
      operationId: deleteMyModelDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: modelDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The definition was deleted
        '404':
          description: No such definition belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The user's collection still references this definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/model-definitions/from/{modelDefinitionId}:
    post:
      tags: [ "PersonalModelDefinitions" ]
      summary: Start customising a shared model definition
      description: >-
        Copies the shared definition, its slots and its wargear options into a personal definition
        owned by the caller, which is then listed alongside the original wherever they see the
        catalogue. Idempotent - returns the existing customisation if they already have one.
      operationId: customiseModelDefinition
      security:
        - bearerAuth: [ ]
      parameters:
        - name: modelDefinitionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The personal copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
        '400':
          description: The given definition is not a shared one
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Model definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/factions:
    get:
      tags: [ "PersonalFactions" ]
      summary: List the signed-in user's own factions
      description: >-
        Both kinds of personal faction: customisations of a shared faction (which carry a
        baseFactionId) and ones the user created from scratch (which do not).
      operationId: getMyFactions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: This user's factions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faction'
    post:
      tags: [ "PersonalFactions" ]
      summary: Create a faction of the user's own
      operationId: createMyFaction
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFactionRequest'
      responses:
        '201':
          description: The created faction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faction'
        '409':
          description: The user already has a faction with that name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/factions/shared:
    get:
      tags: [ "PersonalFactions" ]
      summary: List the shared factions on their own
      description: >-
        GET /api/v1/factions returns the shared factions and the user's own together. This endpoint
        returns the shared ones alone: to offer them for customisation, and to diff a personal copy
        against the faction it was forked from.
      operationId: getSharedFactions
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: The shared factions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faction'
  /api/v1/my/factions/{factionId}:
    put:
      tags: [ "PersonalFactions" ]
      summary: Update one of the user's own factions
      operationId: updateMyFaction
      security:
        - bearerAuth: [ ]
      parameters:
        - name: factionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFactionRequest'
      responses:
        '200':
          description: The updated faction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faction'
        '404':
          description: No such faction belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: The user already has a faction with that name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags: [ "PersonalFactions" ]
      summary: Delete one of the user's own factions, reverting to the shared one if it was a customisation
      operationId: deleteMyFaction
      security:
        - bearerAuth: [ ]
      parameters:
        - name: factionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The faction was deleted
        '404':
          description: No such faction belongs to this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Model definitions still belong to this faction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/my/factions/from/{factionId}:
    post:
      tags: [ "PersonalFactions" ]
      summary: Start customising a shared faction
      description: >-
        Copies the shared faction into one owned by the caller, which is then listed alongside the
        original. Idempotent - returns the existing customisation if they already have one.
      operationId: customiseFaction
      security:
        - bearerAuth: [ ]
      parameters:
        - name: factionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The personal copy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Faction'
        '400':
          description: The given faction is not a shared one
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Faction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
      required: [ message ]
...
    ModelDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          description: >-
            Stable identifier from an imported reference dataset (e.g. battlereadyshelf-data),
            used to match this definition across re-imports. Null for definitions created directly
            in the admin UI rather than via import.
        factionId:
          type: string
          format: uuid
          readOnly: true
          description: The faction this model belongs to, if any.
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            Null for the shared catalogue that admins curate. Set when this is a user's own model
            definition, which only they can see or edit.
        baseModelDefinitionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            The shared definition a personal one was customised from, or null when the user wrote
            it from scratch. Drives the personal-vs-global diff and the revert action.
        name:
          type: string
        description:
          type: string
        version:
          type: integer
          format: int32
          readOnly: true
          description: Incremented every time this model definition is published.
        attachmentSlots:
          readOnly: true
          type: array
          description: >-
            Slots on this model (e.g. "Left Arm", "Right Arm") that can be filled with wargear.
          items:
            $ref: '#/components/schemas/AttachmentSlot'
        wargearOptions:
          readOnly: true
          type: array
          description: >-
            Known wargear options for this model, and which slot(s) each one fills. This is not
            an exhaustive/enforced list - it exists to describe common/default loadouts.
          items:
            $ref: '#/components/schemas/WargearOption'
        hidden:
          type: boolean
          readOnly: true
          description: >-
            Whether the caller has taken this out of their own pickers. Per-user, so hiding a shared
            definition affects nobody else. Only ever true on the /api/v1/my lists, which return
            hidden definitions so they can be shown again; the lists behind pickers omit them. True
            here also when the model's faction is hidden, since hiding a faction hides its models.
      required: [ name ]
    Faction:
      type: object
      description: >-
        Categorises model definitions for catalogue organisation/filtering only - it carries no
        army-building rules (legality, availability, inclusion/exclusion, points, etc.).
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          nullable: true
          description: >-
            Stable identifier used to match this faction across imports. Null for a faction a user
            created for themselves, which the reference dataset knows nothing about.
        name:
          type: string
        parentFactionId:
          type: string
          format: uuid
          nullable: true
          description: Null if this faction has no parent.
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            Null for the shared catalogue. Set on a faction a user added or customised for
            themselves, in which case only they can see it.
        baseFactionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            The shared faction this personal one was forked from, so the app can show how it
            differs and offer to revert. Null both for shared factions and for ones a user invented.
        hidden:
          type: boolean
          readOnly: true
          description: >-
            Whether the caller has taken this out of their own pickers. Per-user, so hiding a shared
            faction affects nobody else. Only ever true on the /api/v1/my lists, which return
            hidden definitions so they can be shown again; the lists behind pickers omit them.
      required: [id, name ]
    AttachmentSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          description: Stable identifier from an imported reference dataset, if this slot was created/matched via import.
        baseAttachmentSlotId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            On a personal model definition, the shared slot this one was customised from. Lets a
            renamed slot read as an edit rather than an unrelated add plus remove.
        name:
          type: string
        type:
          type: string
          description: Stable machine-readable category for this physical attachment slot.
      required: [ name, type ]
    WargearOption:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          description: Stable identifier from an imported reference dataset, if this option's wargear came from one.
        baseWargearOptionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            On a personal model definition, the shared wargear option this one was customised from.
        wargearDefinitionId:
          type: string
          format: uuid
          readOnly: true
          description: The shared wargear definition this option refers to.
        name:
          type: string
          readOnly: true
          description: Resolved from the shared wargear definition; edit it there to rename it everywhere.
        isDefault:
          type: boolean
          description: Whether this option is part of the model's default loadout.
        attachmentSlotIds:
          type: array
          description: >-
            The slot(s) this wargear option fills. An option filling more than one slot (e.g. a
            two-handed weapon) consumes all of those slots at once.
          items:
            type: string
            format: uuid
      required: [ name, isDefault, attachmentSlotIds ]
...
    UpsertModelDefinitionDraftRequest:
      type: object
      description: >-
        Full replacement of a draft's contents. Attachment slots and wargear options that already
        exist in the draft should be sent with their existing draft id so they are updated in
        place; new ones should have a client-generated id (a fresh random UUID) so wargear options
        in the same request can reference them by that id. Any existing slot/option not present in
        the request is removed from the draft.
      properties:
        name:
          type: string
        faction_id:
          type: string
          format: uuid
          description: The faction this model belongs to
        description:
          type: string
        attachmentSlots:
          type: array
          items:
            $ref: '#/components/schemas/UpsertAttachmentSlotDraftRequest'
        wargearOptions:
          type: array
          items:
            $ref: '#/components/schemas/UpsertWargearOptionDraftRequest'
      required: [ name, attachmentSlots, wargearOptions ]
    UpsertAttachmentSlotDraftRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The draft slot id (existing, or a freshly client-generated one for a new slot).
        name:
          type: string
        type:
          type: string
          description: Stable machine-readable category for this physical attachment slot.
      required: [ id, name, type ]
    UpsertWargearOptionDraftRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The draft option id (existing, or a freshly client-generated one for a new option).
        wargearDefinitionId:
          type: string
          format: uuid
          description: >-
            The shared wargear definition to point at. Takes precedence over 'name' when both are
            sent; when omitted, the definition is looked up by name and created if it does not
            exist yet.
        name:
          type: string
          description: >-
            Used to resolve or create the shared wargear definition when 'wargearDefinitionId' is
            not supplied. Renaming here creates/points at a different definition rather than
            renaming the existing one - rename via the wargear definition itself.
        isDefault:
          type: boolean
        attachmentSlotIds:
          type: array
          items:
            type: string
            format: uuid
      required: [ id, name, isDefault, attachmentSlotIds ]
    WargearDefinition:
      type: object
      description: >-
        A canonical piece of wargear shared by every model definition that can be equipped with it,
        so it is named once rather than copied per model. Ids form a single global namespace
        because the reference dataset already reuses them across factions.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        externalId:
          type: string
          readOnly: true
          description: Stable identifier from the reference dataset; null for wargear created by hand.
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            Null for the shared catalogue. Set when a user named wargear the catalogue doesn't
            have while building a personal model definition, in which case only they can see it.
        name:
          type: string
        baseWargearDefinitionId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: >-
            The shared wargear this personal one was forked from, so the app can show how it
            differs and offer to revert. Null both for shared wargear and for wargear a user
            named themselves.
        usageCount:
          type: integer
          format: int32
          readOnly: true
          description: How many published model definitions currently reference this wargear.
        hidden:
          type: boolean
          readOnly: true
          description: >-
            Whether the caller has taken this out of their own pickers. Per-user, so hiding shared
            wargear affects nobody else. Only ever true on the /api/v1/my lists, which return
            hidden definitions so they can be shown again; the lists behind pickers omit them.
      required: [ name ]
...
    UpdateFactionRequest:
      type: object
      description: >-
        The complete proposed state of a faction. An omitted or null parentFactionId proposes
        clearing the parent rather than leaving it alone.
      properties:
        name:
          type: string
        parentFactionId:
          type: string
          format: uuid
          nullable: true
      required: [ name ]
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/UserRole'
        themePreference:
          allOf:
            - $ref: '#/components/schemas/ThemePreference'
      required: [ id, email ]
    UserRole:
      type: string
      enum: [ GUEST, USER, ADMIN, SUPERADMIN ]
    ThemePreference:
      type: string
      description: >-
        A user's preferred UI colour scheme. AUTO follows the device's system preference.
      enum: [ LIGHT, DARK, AUTO ]
...